requirements.txt
Gunicorn
requests
python-dotenv
Flask
Flask-Cors 
google-ai-generativelanguage
google-auth
google-generativeai
json5
youtube_transcript_api=0.6.1